current WIP

+++ jkent_net/repository.py
@@ -15,37 +15,35 @@ class Repository:
     def path(self):
         return self._path
 
-    def exists(self, path, version='HEAD'):
-        try:
-            subprocess.check_call(['/usr/bin/git', '-C', self._path, 'cat-file',
-                '-e', '{}:{}'.format(version, path)], stderr=subprocess.DEVNULL)
-        except subprocess.CalledProcessError:
-            return False
-        return True
-
-    def read(self, path, version='HEAD'):
-        if not version:
-            if not os.path.isfile(os.path.join(self._path, path)):
-                return None
-            try:
-                file = open(os.path.join(self._path, path), 'rb')
-            except FileNotFoundError:
-                return None
-            return file
+    def exists(self, path, version=None):
+        if version == None:
+            return os.path.exists(os.path.join(self.path, path))
         else:
+            try:
+                subprocess.check_call(['/usr/bin/git', '-C', self._path, 'cat-file',
+                    '-e', '{}:{}'.format(version, path)], stderr=subprocess.DEVNULL)
+            except subprocess.CalledProcessError:
+                return False
+            return True
+
+    def open(self, path, mode, version=None):
+        if not version:
+            return open(os.path.join(self._path, path), mode)
+        elif version and 'r' in mode:
             try:
                 data = subprocess.check_output(['/usr/bin/git', '-C',
                     self._path, 'cat-file', '-p',
                     '{}:{}'.format(version, path)], stderr=subprocess.DEVNULL)
             except subprocess.CalledProcessError:
-                return None
+                raise FileNotFoundError
             return BytesIO(data)
+        elif version and 'w' in mode:
+            raise ValueError
 
-    def write(self, path, data):
-        with open(os.path.join(self._path, path), 'wb') as f:
-            f.write(data)
+    def checkout(self, path, version):
+        if version == None:
+            raise ValueError
 
-    def checkout(self, path, version='HEAD'):
         subprocess.check_call(['/usr/bin/git', '-C', self._path, 'checkout',
             version, '--', path])
         subprocess.check_call(['/usr/bin/git', '-C', self._path, 'clean', '-fd',
@@ -59,9 +57,9 @@ class Repository:
             '--allow-empty-message', '-m', message]
         subprocess.check_call(command)
 
-    def history(self, path=None, version='HEAD', num=None):
+    def history(self, path=None, version=None, num=None):
         if not version:
-            return []
+            raise ValueError
 
         try:
             result = subprocess.check_output(['/usr/bin/git', '-C', self._path,
@@ -79,37 +77,51 @@ class Repository:
                     results.append(line)
         return results
 
-    def list(self, path, version='HEAD', recursive=False):
+    def list(self, path, version=None, recursive=False):
+        path = os.path.normpath(path)
+        if not self.isdir(path, version):
+            return ValueError
+        if not path.endswith('/'):
+            path += '/'
         if version:
             try:
-                prefix = os.path.dirname(path)
                 result = subprocess.check_output(['/usr/bin/git', '-C',
                     self._path, 'ls-tree'] + (['-r'] if recursive else []) +
-                    ['--name-only', version, '--', path],
+                    [version, '--', path],
                     stderr=subprocess.DEVNULL)
-                paths = result.rstrip().split(b'\n')
             except subprocess.CalledProcessError:
-                paths = []
+                return []
+
+            paths = []
+            lines = result.decode('utf8').rstrip().split('\n')
+            for line in lines:
+                _, type_, _, filename = line.split(None, 3)
+                if type_ == 'tree':
+                    filename += '/'
+                paths.append(filename)
+
+            paths = list(map(lambda x: x[len(path):], paths))
         else:
-            prefix = os.path.dirname(os.path.join(self._path, path))
+            path = os.path.dirname(os.path.join(self._path, path))
             paths = []
             for root, dirs, files in os.walk(os.path.join(self._path, path)):
                 for dir in dirs:
                     dirname = os.path.normpath(os.path.join(root, dir)) + os.sep
-                    paths.append(dirname.encode('utf8'))
+                    paths.append(dirname)
                 if recursive:
                     dirs[:] = [d for d in dirs if not d.startswith('.')]
                 else:
                     dirs[:] = []
                 for file in files:
                     filename = os.path.normpath(os.path.join(root, file))
-                    paths.append(filename.encode('utf8'))
+                    paths.append(filename)
 
-        paths = list(map(lambda x: x[len(prefix) + 1:], paths))
+            paths = list(map(lambda x: x[len(path) + 1:], paths))
+
         paths.sort()
         return paths
 
-    def isdir(self, path, version='HEAD'):
+    def isdir(self, path, version=None):
         if version:
             try:
                 result = subprocess.check_output(['/usr/bin/git', '-C',
